+++ tests/src/UtilTest.php
@@ -3,7 +3,7 @@
  * Util test.
  *
  * @author    Pronamic <info@pronamic.eu>
- * @copyright 2005-2019 Pronamic
+ * @copyright 2005-2020 Pronamic
  * @license   GPL-3.0-or-later
  * @package   Pronamic\WordPress\Pay\Gateways\Sisow
  */
@@ -15,7 +15,7 @@ use WP_UnitTestCase;
 /**
  * Title: Sisow util test
  * Description:
- * Copyright: 2005-2019 Pronamic
+ * Copyright: 2005-2020 Pronamic
  * Company: Pronamic
  *
  * @author  Remco Tolsma