+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## [Unreleased][unreleased]
 -
 
+## [2.1.0] - 2020-03-19
+- Catch exception in gateway instead of client.
+- Extend from AbstractGatewayIntegration class.
+
 ## [2.0.4] - 2019-12-22
 - Added support for new `pronamic_pay_return_should_redirect` filter for notify and callback processing.
 - Added URL to manual in gateway settings.
@@ -94,7 +98,8 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## 1.0.0 - 2014-12-12
 - First release.
 
-[unreleased]: https://github.com/wp-pay-gateways/sisow/compare/2.0.4...HEAD
+[unreleased]: https://github.com/wp-pay-gateways/sisow/compare/2.1.0...HEAD
+[2.1.0]: https://github.com/wp-pay-gateways/sisow/compare/2.0.4...2.1.0
 [2.0.4]: https://github.com/wp-pay-gateways/sisow/compare/2.0.3...2.0.4
 [2.0.3]: https://github.com/wp-pay-gateways/sisow/compare/2.0.2...2.0.3
 [2.0.2]: https://github.com/wp-pay-gateways/sisow/compare/2.0.1...2.0.2